Migrate reset PIN related tests to a separate test class
